updates to util and main
