+++ data_handling/munge_ansirh.py
@@ -8,6 +8,7 @@ import pandas as pd
 import json
 from util.constants import STANDARD_ENCODING
 from data_handling.clean_ansirh import clean_ansirh
+from util.functions import translate_code_to_state
 
 
 def main():
@@ -71,27 +72,6 @@ def split_by_state(rows):
             state_dict[full_state].append(row)
 
     return state_dict
-
-
-def translate_code_to_state(state_abr):
-    """
-    Turns two-letter state code into the full state name.
-
-    Inputs:
-        state_abr (str): two letter state abbreviation
-
-    Returns (str):
-        Full state name.
-    """
-    # TODO: Will refactor with csv in the near future
-    # Read in state abreviation data
-    file_name = "./data/state_abbreviations.csv"
-    state = pd.read_csv(file_name)
-
-    # Convert to full state name
-    state_name = state["state"][state["code"] == state_abr.upper()].values[0]
-
-    return state_name
 
 
 def split_by_zip(state_dict):
